added semi-analytics constant files
